* Added BlockEntity. * Level "block" is finished.
I noticed that sometimes there is a bug where a lemming gets stuck in a corner.  It happens very rarely.  If anyone can find a reliable way to reproduce, please let me know.

I also notice that we are starting to have performance problems (fps ~ 20 when it should be 30).  Are you seeing this on Windows too?